esp32s3-usbserial: split usb
